initial commit for SpringSecurity2 resolved circular dependency problem.

+++ src/main/java/com/CinemaHub/CinemaHub/Repository/SeatRepo.java
@@ -1,7 +1,6 @@
 package com.CinemaHub.CinemaHub.Repository;
 
 
-
 import com.CinemaHub.CinemaHub.Entity.Seat;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
@@ -12,12 +11,11 @@ import org.springframework.stereotype.Repository;
 import javax.transaction.Transactional;
 import java.util.List;
 
-
 @Repository
 public interface SeatRepo extends JpaRepository<Seat,Long> {
 
     @Query("SELECT s FROM Seat s WHERE s.movieShow.showId = :showId AND s.isReserved = FALSE")
-    List<Seat> findNonReservedSeatsByShowId(@Param("showId") Long showId);
+    List<Seat> findNonReservedSeatsByShowId(@Param("showId") Integer  showId);
 
     @Modifying
     @Transactional
@@ -30,7 +28,9 @@ public interface SeatRepo extends JpaRepository<Seat,Long> {
     void updateSeatReservationStatus2(List<Integer> seatNos, List<Integer> showIds);
 
 
-    List<Seat> findBySeatNoInAndMovieShow_ShowId(List<Integer> seatNumbers, Long showId);
+    List<Seat> findBySeatNoInAndMovieShow_ShowId(List<Integer> seatNumbers, Integer showId);
 }
+
+
